mk_sdcard_image: Clean up not needed code
Remove not used parameters of functions.

Remove not needed global variables DOM*_PRESENT.
This variable is set only if parameters of partition are defined.
So if DOM*_START is set, this mean that partition is present.

Remove creation of not needed loop devices. Now we create one loop
device for image output, and one for android's disk.
For now android's loop device is created/deleted two times: for
partitioning, and for writing images. To fix this we need to
significantly rework script, so effort is to big for small reward.

Signed-off-by: Ruslan Shymkevych <ruslan_shymkevych@epam.com>
Reviewed-by: Oleksandr Andrushchenko <oleksandr_andrushchenko@epam.com>