adas/device-tree: Add interrupt parent for out of "/soc" devices
Some devices (which are to be passed through) lay out of device
tree's "soc" section, thus their interrupt-parent property is not
set to &gic making Xen skip those for vGIC of Dom0. So, we have to
manually set interrupt-parent to "gic" for such nodes.
DomD's device tree then needs to set corresponding interrupt-parent to
PHANDLE_GIC as it is done for "SOC" section of DomD.

Signed-off-by: Oleksandr Andrushchenko <oleksandr_andrushchenko@epam.com>
Signed-off-by: Oleksandr Tyshchenko <oleksandr_tyshchenko@epam.com>
Reviewed-by: Viktor Mitin <viktor_mitin@epam.com>