Use dedicated PulseAudio config for SK+KF
We used to use udev-detect module to find availabe cards on SK+KF, but this
results in two issues:
1) not predictible order of soundcards
2) sometimes pcm3168a is detected and loaded to slow, and system.pa
script is not able to set it as default.

Patch solves these issues by removing udev-detect and directly loading
alsa-card modules for each sound card. At this point we need to separate
system.pa for SK+KF and other cards, as far as we have only ak4613 on
all boards (salvator-x, -xs, starterkit) exept starterkit with kingfisher
where we have pcm3168a, ak4613, radio and bt. Last two are not used
for now so alsa-card module is not loaded for them.
Also remove 'set-default-sink' for not SK+KF boards.

Suggested-by: Volodymyr Babchuk <volodymyr_babchuk@epam.com>
Signed-off-by: Ruslan Shymkevych <ruslan_shymkevych@epam.com>