domd: weston: Support ivi-id-agent
Port the patches for supporting ivi-id-agent added by previous commit.

According to:
https://github.com/mtey/weston/commits/xdg_support_ivi_id_agent

---

The main reason of these changes is an ability to use existing desktop applications
without alternation (e.g. gstreamer sink plugin).

Without this support the gstreamer sink plugin complains about:
Failed to create GstWlDisplay: 'Could not bind to wl_shell. Either it is
not implemented in the compositor, or the implemented version doesn't match'

Signed-off-by: Oleksandr Tyshchenko <oleksandr_tyshchenko@epam.com>