Cleaning of rcar-proprietary-graphic.bb
Motivation
The rcar-proprietary-graphic.bb recipes (domd/domu) has a lot
of common code. To mitigate the risk of the error during
the modification of the common part, it has been decided
to move the common functionality in the separate file.
Also, it is required to adjust the existed recipes code
with .../gles-module/rcar-proprietary-graphic.inc

The scope
1. clean up according to issue :
https://github.com/xen-troops/meta-xt-prod-gen3-test/issues/164 (part 1)

The idea was to adjust the code according to teh sample from
https://github.com/renesas-rcar/meta-renesas/blob/thud-Yocto-v3.21.0/
meta-rcar-gen3/recipes-graphics/gles-module/gles-user-module.bb

Main changes:
- Supported machines : r8a7795, r8a7796, r8a77965
- The SCR_URI is costomized just for supported machines :
 SRC_URI_r8a7795
 SRC_URI_r8a7796
 SRC_URI_r8a77965

2. The common code from rcar-proprietary-graphic/rcar-proprietary-graphic.bb
(domd and domu of prod-devel) is moved into
recipes-domx/meta-xt-images-vgpu/recipes-graphics/
gles-module/rcar-proprietary-graphic.inc

Signed-off-by: Usyk Ihor <ihor_usyk@epam.com>